Correct sha256sum and final check
Move to target directory first, checksum file read with "-c" generally
contains only the file name, and sha256sum assumes it relative to the
working directory instead of the directory the checksum file is in.

Fail if the target file does not exist, it would not be a sign of
success.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>